+++ test/cli/agent_cli_inbox_drain_timeout_test.dart
@@ -12,6 +12,9 @@ class _WedgedHubPlugin implements FahPlugin {
   String get name => 'wedged_hub';
 
   @override
+  Future<void> dispose() async {}
+
+  @override
   void register(PluginContext context) {
     context.registerExternalInbox(
       ExternalInbox(drain: () => Completer<List<AgentMessage>>().future),